requirements.txt
chromadb>=0.5.0
discord.py>=2.4.0
openai>=1.40.0
python-dotenv>=1.0.1
tiktoken>=0.7.0
